build: garden support (#511)
Add garden project and module for the service. It spins up a kubernetes environment with all the dependencies for the service.
I created two new dockerfiles one for development which will be used for hot reload and the other for production builds using garden. We may want to consider using the multistage for real production workloads as well